@@ -31,7 +31,5 @@ public interface OnTrimVideoListener {
 
     void getResult(final Uri uri);
 
-    void cancelAction();
-
-    void onError(final String message);
+    void onError(final Throwable throwable);
 }
