@@ -24,8 +24,9 @@
 package life.knowledge4.videotrimmer.utils;
 
 import android.net.Uri;
-import android.support.annotation.NonNull;
 import android.util.Log;
+
+import androidx.annotation.NonNull;
 
 import com.coremedia.iso.boxes.Container;
 import com.googlecode.mp4parser.FileDataSourceViaHeapImpl;
@@ -60,7 +61,6 @@ public class TrimVideoUtils {
         final String filePath = dst + fileName;
 
         File file = new File(filePath);
-        file.getParentFile().mkdirs();
         Log.d(TAG, "Generated file path " + filePath);
         genVideoUsingMp4Parser(src, file, startMs, endMs, callback);
     }
@@ -122,8 +122,6 @@ public class TrimVideoUtils {
             }
             movie.addTrack(new AppendTrack(new CroppedTrack(track, startSample1, endSample1)));
         }
-
-        dst.getParentFile().mkdirs();
 
         if (!dst.exists()) {
             dst.createNewFile();
